Breaking the DI logic to an extention method so all DI related logic is now in one place and not duplicated in the code base.  Also fixing the property binding which appears to have been set in correctly.
