+++ Powershell5.DI/Powershell5.DI/DI/IShouldInject.cs
@@ -1,0 +1,7 @@
+﻿namespace UTMO.Powershell5.DI.DI
+{
+    public interface IShouldInject
+    {
+        string Name { get; set; }
+    }
+}
